Fix `z.object` bug with cyclic references
